Add files via upload
Exciting news! I recently participated in my first hackathon, and I'm thrilled to share that my app garnered recognition by being nominated in the MTN Business App of the Year Awards 2023. This experience was not only a fantastic learning opportunity but also a testament to the hard work and innovation poured into the development process. I'm grateful for the acknowledgment and motivated to continue exploring new horizons in the dynamic world of app development.